Refactor transform functions, pass transforms as an argument to dataflux dataset constructor
